+++ src/main/java/com/keunsori/keunsoriserver/reservation/domain/vo/Session.java
@@ -1,4 +1,4 @@
-package com.keunsori.keunsoriserver.reservation.vo;
+package com.keunsori.keunsoriserver.reservation.domain.vo;
 
 public enum Session {
     VOCAL, KEYBOARD, DRUM, GUITAR, BASS, ALL
