+++ src/main/java/com/keunsori/keunsoriserver/reservation/dto/ReservationCreateRequest.java
@@ -1,9 +1,9 @@
 package com.keunsori.keunsoriserver.reservation.dto;
 
 import com.keunsori.keunsoriserver.member.Member;
-import com.keunsori.keunsoriserver.reservation.Reservation;
-import com.keunsori.keunsoriserver.reservation.vo.ReservationType;
-import com.keunsori.keunsoriserver.reservation.vo.Session;
+import com.keunsori.keunsoriserver.reservation.domain.Reservation;
+import com.keunsori.keunsoriserver.reservation.domain.vo.ReservationType;
+import com.keunsori.keunsoriserver.reservation.domain.vo.Session;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.sql.Time;
